added back ability to vote, save, and mark posts as read
